Bug fixed

+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/metrics/internal/NpathBaseVisitor.java
@@ -97,7 +97,8 @@ public class NpathBaseVisitor extends JavaVisitorBase<Void, BigInteger> {
 
         int boolComp = CycloVisitor.booleanExpressionComplexity(node.getCondition());
         BigInteger nPathBody = node.getBody().acceptVisitor(this, data);
-        return nPathBody.add(BigInteger.valueOf(boolComp + 1));
+        // return nPathBody.add(BigInteger.valueOf(boolComp + 1));
+        return nPathBody.add(BigInteger.valueOf((long) boolComp + 1));
     }
 
 
@@ -107,7 +108,8 @@ public class NpathBaseVisitor extends JavaVisitorBase<Void, BigInteger> {
 
         int boolComp = CycloVisitor.booleanExpressionComplexity(node.getCondition());
         BigInteger nPathBody = node.getBody().acceptVisitor(this, data);
-        return nPathBody.add(BigInteger.valueOf(boolComp + 1));
+        // return nPathBody.add(BigInteger.valueOf(boolComp + 1));
+        return nPathBody.add(BigInteger.valueOf((long) boolComp + 1));
     }
 
 
@@ -117,7 +119,8 @@ public class NpathBaseVisitor extends JavaVisitorBase<Void, BigInteger> {
 
         int boolComp = CycloVisitor.booleanExpressionComplexity(node.getCondition());
         BigInteger nPathBody = node.getBody().acceptVisitor(this, data);
-        return nPathBody.add(BigInteger.valueOf(boolComp + 1));
+        // return nPathBody.add(BigInteger.valueOf(boolComp + 1));
+        return nPathBody.add(BigInteger.valueOf((long) boolComp + 1));
     }
 
     @Override
